Test statically linked clang outside of Alpine Linux
We should verify that we can use the toolchain produced in Alpine Linux
on other distributions as part of the continuous integration pipeline.

We download the toolchain out of the container, mount it into the
containers that we want to test it in, and test dynamically and
statically linking hello.{c,cpp} with musl. Eventually, glibc should be
tested but there are issues during the link phase.

Initially, the only container that is being tested is Fedora. However, I
intentionally wrote this in such a way that we should be able to use
this across multiple distributions and test more than just
hello.{c,cpp}.

This was tested on a self-hosted GitHub Actions runner:

https://github.com/nathanchance/containers/actions/runs/2387476728

Signed-off-by: Nathan Chancellor <nathan@kernel.org>