Upload toolchain tarball when build is successful
Currently, the toolchain is only available through a Docker image, which
is quite large and only uploaded from our organization. For testing, it
would be useful to have some way to build in CI and pull down the result
without having to upload a new image to the container registry.

Take advantage of the upload-artifact action, which allows us to create
then upload the toolchain tarball so it can be downloaded after a
successful run. Because we are a public repository, we do not have any
usage limits with storage.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>