ci/install-vm.py: Explicitly connect to qemu:///system
In order for virtual machines to automatically restart across reboots,
they need to be "system" virtual machines, not "session" virtual
machines.

This script is designed to be run without root, which means that by
default, libvirt will use qemu:///session. This can be overridden in the
environment with the LIBVIRT_DEFAULT_URI variable, which I would
recommend people set, but make sure the script works without that
variable by using the '--connect' parameter.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>