github: Make tag specific to architecture
Currently, we share a tag between arm64 and x86_64, which does not work
without a manifest or using 'docker buildx' with the '--platform' flag
containing multiple builds.

A manifest would be useful but it still requires separate architecture
specific tags anyways.

We cannot rely on the '--platform' flag because building arm64 on x86_64
would run through qemu-user-static, which is slower than doing the build
on a native arm64 host.

To solve this, add the architecture to the tag so that we can generate
unique tags that will work for each architecture.

Use "aarch64" instead of "arm64" for the arch value so that we can
easily use 'uname -m' in a future Makefile patch.

Link: https://docs.docker.com/engine/reference/commandline/manifest/
Signed-off-by: Nathan Chancellor <nathan@kernel.org>