github: build-test-llvm-project: Add build-args
We need this now after commit dcc15cd ("github: Make tag specific to
architecture"). Unfortunately, I don't see a clear way to pass in
build-args conditionally to docker/build-push-action, so we provide a
dummy list of arguments that can be used when no build arguments are
needed (i.e., for epoch one).

Signed-off-by: Nathan Chancellor <nathan@kernel.org>